requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.3
PyYAML==6.0.2
openai==1.75.0
